Revert the refactoring code to avoid execlusion in public api check
