+++ pom.xml
@@ -522,9 +522,6 @@
                 change to fix a integer overflow issue.
                 TODO: remove this after Parquet 1.13 release -->
               <exclude>org.apache.parquet.column.values.dictionary.DictionaryValuesWriter#dictionaryByteSize</exclude>
-              <exclude>org.apache.parquet.hadoop.util.ColumnMasker.MaskMode</exclude>
-              <exclude>org.apache.parquet.hadoop.util.CompressionConverter</exclude>
-              <exclude>org.apache.parquet.hadoop.util.ColumnMasker</exclude>
             </excludes>
           </parameter>
         </configuration>
